transformations: add lowering from memref_stream generic to loops
